requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
numpy==2.1.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
